refactor slack actions into seperate functions
